OSIS-3523 : Move all files into webservices
