+++ education_group/api/url_v1.py
@@ -25,16 +25,10 @@
 ##############################################################################
 from django.conf.urls import url
 
-from education_group.api.views.general_information import GeneralInformation
 from education_group.api.views.training import TrainingList, TrainingDetail
 
 app_name = "education_group"
 urlpatterns = [
     url(r'^trainings/$', TrainingList.as_view(), name=TrainingList.name),
     url(r'^trainings/(?P<uuid>[0-9a-f-]+)$', TrainingDetail.as_view(), name=TrainingDetail.name),
-    url(
-        r'^trainings/(?P<year>[\d]{4})/(?P<language>[\w]{2})/(?P<acronym>[\w]+)$',
-        GeneralInformation.as_view(),
-        name=GeneralInformation.name
-    )
 ]
