+++ webservices/urls.py
@@ -30,6 +30,7 @@ from django.conf import settings
 from django.conf.urls import url, include
 
 from webservices.api.views.auth_token import AuthToken
+from webservices.api.views.general_information import GeneralInformation
 from webservices.views import ws_catalog_offer, ws_catalog_common_offer, ws_catalog_common_admission_condition, \
     ws_catalog_offer_v02
 
@@ -72,5 +73,8 @@ urlpatterns = [
     url('^v0.2/catalog/offer/(?P<year>[0-9]{4})/(?P<language>[a-zA-Z]{2})/(?P<acronym>[a-zA-Z0-9]+)$',
         ws_catalog_offer_v02,
         name='v0.2-ws_catalog_offer'),
+    url('^v0.3/catalog/offer/(?P<year>[0-9]{4})/(?P<language>[a-zA-Z]{2})/(?P<acronym>[a-zA-Z0-9]+)$',
+        GeneralInformation.as_view(),
+        name=GeneralInformation.name),
     url(r'^v1/', include(url_api_v1)),
 ]
